moved queries into sub-directory
